move utils/metric.py to evaluators/utils.py
