+++ src/kwja/evaluators/typo_module_metric.py
@@ -5,7 +5,7 @@ from Levenshtein import opcodes
 from torchmetrics import Metric
 
 from kwja.datamodule.datasets.typo_dataset import TypoDataset
-from kwja.utils.metric import unique
+from kwja.evaluators.utils import unique
 from kwja.utils.typo_module_writer import apply_edit_operations, convert_predictions_into_typo_corr_op_tags
